requirements.txt
web.py
librosa
pyyaml
unmix
json
pafy
youtube_dl